Added ability to Orchestrator to monitor a crank file in a Git repository and respond to changes to it.
Monitoring is achieved via polling the repository at a configurable interval. The repository can be either local (on same machine) or remote. If a repository is remote, Orchestrator first creates a bare clone of it in SLING_HOME, and upon each poll does "git fetch" first. Each time the monitored file is updated, it is copied inside SLING_HOME with the name C<timestamp>.crank.txt.

Using JGit library which had to be embedded together with dependencies because one of them is not a valid bundle, but yet listed as "Require-Bundle" in JGit's bundle (making it impossible to replace it with a different bundle exporting the necessary packages).