Added the sample bundle that contains a service and a script.
Added the project of the sample bundle, of which two versions, in the form of JAR bundles, were provided before in the "resources" folder. The default version of the bundle is 0.0.1. To build the 0.0.2 version, with relevant source code modifications, the Maven profile "2" must be used (e.g. "mvn -P2 clean install") - source code is transformed appropriately before being compiled.